AUV/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.1
dm-tree==0.1.8
EasyProcess==1.1
entrypoint2==1.1
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.9.2
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
huggingface-hub==0.17.3
idna==3.4
keras==2.14.0
keras-core==0.1.7
keras-nlp==0.6.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.2.0
MouseInfo==0.1.3
mss==9.0.1
namex==0.0.7
numpy==1.26.1
oauthlib==3.2.2
onnx==1.14.1
onnxconverter-common==1.13.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pyserial==3.5
python-dateutil==2.8.2
python-statemachine==2.1.2
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
rubicon-objc==0.4.7
safetensors==0.4.0
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-cpu==2.14.0
tensorflow-estimator==2.14.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-text==2.14.0
termcolor==2.3.0
tf2onnx==1.15.1
tokenizers==0.14.1
tqdm==4.66.1
transformers==4.34.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.0
wrapt==1.14.1
